# src/main/java/com/hekreminder/dto/ReminderListRequest.java
package com.hekreminder.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;

@Getter
@NoArgsConstructor
@AllArgsConstructor
public class ReminderListRequest {

    @NotBlank
    private String name;

    private String color;

    private String icon;
}